requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
llama-index
llama-index-llms-azure-openai
llama-index-storage-chat-store-postgres
jinja2
python-multipart